Added confirmation on delete

+++ src/lib/instances/delete/delete.sh
@@ -56,44 +56,50 @@ CAD=$(pwd)
 LINKNAME="${HOME}/.doil/$INSTANCE"
 if [ -h "${LINKNAME}" ]
 then
-  # Pipe output to null if needed
-  if [[ ${QUIET} == TRUE ]]
+
+  read -p "Please confirm that you want to delete ${INSTANCE} [yN]: " CONFIRM
+  if [[ ${CONFIRM} == "y" ]]
   then
-    exec >>/var/log/doil.log 2>&1
+
+    # Pipe output to null if needed
+    if [[ ${QUIET} == TRUE ]]
+    then
+      exec >>/var/log/doil.log 2>&1
+    fi
+
+    doil_send_log "Deleting instance"
+
+    # check saltmain
+    doil system:salt start --quiet
+
+    # check proxy server
+    doil system:proxy start --quiet
+
+    # set machine inactive
+    doil down ${INSTANCE} --quiet
+
+    # remove directory
+    the_path=$(readlink ${LINKNAME})
+    sudo rm -rf $the_path
+
+    # remove link
+    rm -f "${HOME}/.doil/$INSTANCE"
+
+    # remove key
+    docker exec -ti saltmain bash -c "echo 'y' | salt-key -d ${INSTANCE}.local"
+
+    # remove proxy
+    if [ -f "/usr/local/lib/doil/tpl/proxy/conf/sites/${INSTANCE}.conf" ]
+    then
+      rm "/usr/local/lib/doil/tpl/proxy/conf/sites/${INSTANCE}.conf"
+    fi
+    doil system:proxy reload --quiet
+
+    # remove docker image
+    docker rmi $(docker images "doil/${INSTANCE}" -a -q)
+
+    doil_send_log "Instance deleted"
   fi
-
-  doil_send_log "Deleting instance"
-
-  # check saltmain
-  doil system:salt start --quiet
-
-  # check proxy server
-  doil system:proxy start --quiet
-
-  # set machine inactive
-  doil down ${INSTANCE} --quiet
-
-  # remove directory
-  the_path=$(readlink ${LINKNAME})
-  sudo rm -rf $the_path
-
-  # remove link
-  rm -f "${HOME}/.doil/$INSTANCE"
-
-  # remove key
-  docker exec -ti saltmain bash -c "echo 'y' | salt-key -d ${INSTANCE}.local"
-
-  # remove proxy
-  if [ -f "/usr/local/lib/doil/tpl/proxy/conf/sites/${INSTANCE}.conf" ]
-  then
-    rm "/usr/local/lib/doil/tpl/proxy/conf/sites/${INSTANCE}.conf"
-  fi
-  doil system:proxy reload --quiet
-
-  # remove docker image
-  docker rmi $(docker images "doil/${INSTANCE}" -a -q)
-
-  doil_send_log "Instance deleted"
 else
   echo -e "\033[1mERROR:\033[0m"
   echo -e "\tInstance not found!"
